feat: Add HTML file for callbacks with event listeners
Add `index-part1-button-clicked.html` file that contains the HTML structure
for implementing callbacks with event listeners. The file includes a button
element with the id "clickMe" and a JavaScript function `buttonClicked()`
that logs "Button clicked" to the console when the button is clicked.

feat: Add HTML file for callbacks with timeouts

Add `index-part2-timeout-clicked.html` file that contains the HTML structure
for implementing callbacks with timeouts. The file includes a button element
with the id "setTimeout" and a JavaScript function `timeoutClicked()` that
logs "Timeout clicked" to the console when the button is clicked. It also
uses `setTimeout()` to display the message "This message is shown 3 seconds
after timeout clicked" after a delay of 3 seconds.

feat: Add HTML file for callbacks with intervals

Add `index-part3-interval-clicked.html` file that contains